Revert "refactor moderations api usage"
This reverts commit b076d3174ca14251ee3fc28bd4744e25685c1e5f.
